Moved joint state connector to RobotConnectors file
