+++ src/lib/RobotConnectors/CMakeLists.txt
@@ -4,5 +4,9 @@ set (HDR RobotConnectors.h)
 set (SRC RobotConnectors.cpp)
 
 add_library(${LIBRARY_NAME} ${HDR} ${SRC})
-target_link_libraries(${LIBRARY_NAME} PRIVATE YARP::YARP_os ReadOnlyRemoteControlBoardLib)
+target_link_libraries(${LIBRARY_NAME} PRIVATE YARP::YARP_os
+                                              YARP::YARP_rosmsg
+                                              YARP::YARP_dev
+                                              ReadOnlyRemoteControlBoardLib)
+
 target_include_directories(${LIBRARY_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
