+++ src/modules/yarprobotstatepublisher/CMakeLists.txt
@@ -13,7 +13,6 @@ target_link_libraries(yarprobotstatepublisher
     YARP::YARP_init
     YARP::YARP_os
     YARP::YARP_dev
-    YARP::YARP_rosmsg
     iDynTree::idyntree-high-level
     iDynTree::idyntree-yarp
     idyntree-yarp-robot-connectors)
