+++ src/modules/yarprobotstatepublisher/include/robotstatepublisher.h
@@ -22,8 +22,6 @@
 #include <iDynTree/Core/VectorDynSize.h>
 #include <iDynTree/KinDynComputations.h>
 
-#include <yarp/rosmsg/sensor_msgs/JointState.h>
-
 #include <RobotConnectors.h>
 
 #include <mutex>
@@ -35,63 +33,6 @@
 /****************************************************************/
 namespace idyntree_yarp_tools {
 
-class JointStateConnector: public BasicConnector
-{
-private:
-
-    class JointStateSubscriber: public yarp::os::Subscriber<yarp::rosmsg::sensor_msgs::JointState>
-    {
-    private:
-        // Mutex protecting the method across the different threads
-        std::mutex m_mutex;
-
-        JointStateConnector* m_connector{nullptr};
-
-    public:
-        JointStateSubscriber();
-
-        ~JointStateSubscriber();
-
-        JointStateSubscriber(const JointStateSubscriber&) = delete;
-        JointStateSubscriber(JointStateSubscriber&&) = delete;
-        JointStateSubscriber& operator=(const JointStateSubscriber&) = delete;
-        JointStateSubscriber& operator=(JointStateSubscriber&&) = delete;
-
-        void attach(JointStateConnector* connector);
-
-        virtual void onRead(yarp::rosmsg::sensor_msgs::JointState &v) override;
-    };
-
-    // Mutex protecting the method across the different threads
-    std::mutex m_mutex;
-    std::mutex m_callbackMutex;
-
-    std::function<void()> m_callback;
-
-    // /JointState topic scruscriber
-    std::unique_ptr<yarp::os::Node> m_rosNode{nullptr};
-    std::unique_ptr<JointStateSubscriber> m_subscriber{nullptr};
-    std::string m_jointStatesTopicName;
-    std::string m_namePrefix;
-    std::unordered_map<std::string, size_t> m_nameToIndexMap;
-    std::atomic<bool> m_connected{false};
-
-    void onRead(yarp::rosmsg::sensor_msgs::JointState &v);
-
-public:
-    void setCallback(std::function<void()> callback);
-
-    bool configure(const yarp::os::Searchable &inputConf, const iDynTree::Model& fullModel, std::shared_ptr<BasicInfo> basicInfo);
-
-    virtual bool connectToRobot() override;
-
-    virtual bool getJointValues(iDynTree::VectorDynSize& jointValuesInRad) override;
-
-    virtual void close() override;
-};
-
-
-/****************************************************************/
 class YARPRobotStatePublisherModule : public yarp::os::RFModule
 {
     std::atomic<double> m_period;
